+++ src/Ethernet_HTTPClient/Ethernet_URLEncoder.h
@@ -7,40 +7,21 @@
    Based on and modified from ESP8266 https://github.com/esp8266/Arduino/releases
    Built by Khoi Hoang https://github.com/khoih-prog/EthernetWebServer
    Licensed under MIT license
-   Version: 1.7.0
 
    Original author:
    @file       Esp8266WebServer.h
    @author     Ivan Grokhotkov
+   
+   Version: 1.7.1
 
    Version Modified By   Date      Comments
    ------- -----------  ---------- -----------
     1.0.0   K Hoang      13/02/2020 Initial coding for Arduino Mega, Teensy, etc to support Ethernetx libraries
-    1.0.1   K Hoang      20/02/2020 Add support to lambda functions
-    1.0.2   K Hoang      20/02/2020 Add support to UIPEthernet library for ENC28J60
-    1.0.3   K Hoang      23/02/2020 Add support to SAM DUE / SAMD21 boards
-    1.0.4   K Hoang      16/04/2020 Add support to SAMD51 boards
-    1.0.5   K Hoang      24/04/2020 Add support to nRF52 boards, such as AdaFruit Feather nRF52832, nRF52840 Express, BlueFruit Sense, 
-                                    Itsy-Bitsy nRF52840 Express, Metro nRF52840 Express, NINA_B302_ublox, etc. 
-                                    More Custom Ethernet libraries supported such as Ethernet2, Ethernet3, EthernetLarge
-    1.0.6   K Hoang      27/04/2020 Add W5x00 support to ESP32/ESP8266 boards
-    1.0.7   K Hoang      30/04/2020 Add ENC28J60 support to ESP32/ESP8266 boards  
-    1.0.8   K Hoang      12/05/2020 Fix W5x00 support for ESP8266 boards.
-    1.0.9   K Hoang      15/05/2020 Add EthernetWrapper.h for easier W5x00 support as well as more Ethernet libs in the future.
-    1.0.10  K Hoang      21/07/2020 Fix bug not closing client and releasing socket.
-    1.0.11  K Hoang      25/07/2020 Add support to Seeeduino SAMD21/SAMD51 boards. Restructure examples.
-    1.0.12  K Hoang      15/09/2020 Add support to new EthernetENC library for ENC28J60. Add debug feature.
-    1.0.13  K Hoang      24/09/2020 Restore support to PROGMEM-related commands, such as sendContent_P() and send_P()
-    1.1.0   K Hoang      17/11/2020 Add basic HTTP and WebSockets Client by merging ArduinoHttpClient
-    1.2.0   K Hoang      04/12/2020 Add support to NativeEthernet Library for Teensy 4.1
-    1.2.1   K Hoang      26/12/2020 Suppress all possible compiler warnings
-    1.3.0   K Hoang      27/01/2021 Add WebServer feature to serve from LittleFS/SPIFFS for ESP32/ESP8266 with examples
-    1.3.1   K Hoang      29/04/2021 Add SimpleWebServer_NativeEthernet and delete AdvancedWebServer_NativeEthernet example
-    1.4.0   K Hoang      13/05/2021 Add support to RP2040-based boards using Arduino mbed_rp2040 core
-    1.5.0   K Hoang      15/05/2021 Add support to RP2040-based boards using Arduino-pico rp2040 core
+    ...
     1.6.0   K Hoang      04/09/2021 Add support to QNEthernet Library for Teensy 4.1
     1.7.0   K Hoang      09/09/2021 Add support to Portenta H7 Ethernet
- *****************************************************************************************************************************/
+    1.7.1   K Hoang      04/10/2021 Change option for PIO `lib_compat_mode` from default `soft` to `strict`. Update Packages Patches
+ *************************************************************************************************************************************/
  
 // Library to simplify HTTP fetching on Arduino
 // (c) Copyright Arduino. 2019